Working with unittests
