+++ application/__init__.py
@@ -1,10 +1,42 @@
 from sanic import Blueprint
 from sanic import Sanic
+from lemkpg.constants import GET_ALL_COLUMNS
+from sanic.exceptions import Unauthorized
 
 from core_api.resources import UrlsView, UrlView, RegisterView, AuthView, RedirectView, DemoView
-
+from core_api.config import USER_TABLE, USER_COLUMNS, db_conn, test_db_conn
+from core_api.utils import response_converter
+import contextvars
 
 app = Sanic("app")
+
+
+@app.middleware('request')
+async def check_authorization_and_add_user_to_request(request):
+    registration_and_auth_paths = ["/api/v1/auth", "/api/v1/register"]
+    api_path = "/api/v1/"
+    if not request.token and request.path not in registration_and_auth_paths and api_path in request.path:
+        raise Unauthorized("Authorization should be defined in request headers")
+    else:
+        result = await db_conn.get(USER_TABLE, GET_ALL_COLUMNS,
+                                   conditions_list=[("token", "=", request.token, None)])
+        if not result and request.path not in registration_and_auth_paths and api_path in request.path:
+            raise Unauthorized("Authorization with Token should be defined in request headers")
+        if result and request.path not in registration_and_auth_paths and api_path in request.path:
+            user = response_converter(result, USER_COLUMNS, ("password",))[0]
+            request["user"] = user
+
+
+@app.middleware('request')
+async def check_run_or_test_and_return_db_conn(request):
+    ctx = contextvars.copy_context()
+    context_values = list(ctx.values())
+    if "default_context" in context_values:
+        request["db_conn"] = db_conn
+    if "test_context" in context_values:
+        request["db_conn"] = test_db_conn
+    else:
+        request["db_conn"] = db_conn
 
 
 def create_api():
